// src/types/smartContractDetailResponse.ts
/**
 * Kriptonio Private API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: private
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import type { WalletResponse } from './walletResponse';
import type { SmartContractDeploymentResponse } from './smartContractDeploymentResponse';
import type { SmartContractTemplateResponse } from './smartContractTemplateResponse';
import type { BlockchainResponse } from './blockchainResponse';
import type { SmartContractType } from './smartContractType';

export interface SmartContractDetailResponse {
  id: string;
  title: string;
  compilerVersion: string;
  optimizerEnabled: boolean;
  optimizerRuns: number;
  template: SmartContractTemplateResponse;
  templateInput: any | null;
  type: SmartContractType;
  abi: string;
  blockchain: BlockchainResponse;
  wallet: WalletResponse;
  deployment?: SmartContractDeploymentResponse;
  createdAt: string;
  paymasterEndpointId?: string | null;
}
export namespace SmartContractDetailResponse {}